+++ src/methods/cytonorm_no_controls_to_mid/script.R
@@ -21,7 +21,16 @@ meta <- list(
 
 source(paste0(meta$resources_dir, "/anndata_to_fcs.R"))
 
-tmp_path <- meta[["temp_dir"]]
+# only for HPC, the idea is if running on HPC, use a temp dir set in the env variable
+tmp_path <- Sys.getenv("HPC_VIASH_META_TEMP_DIR")
+if (tmp_path != "") {
+  # Environment variable is set, use it
+  print(paste0("Using HPC temp dir from env: ", tmp_path))
+} else {
+  # Environment variable not set, use meta
+  tmp_path <- meta[["temp_dir"]]
+  print(paste0("Using meta temp dir: ", tmp_path))
+}
 
 cat("Reading input files\n")
 adata <- anndata::read_h5ad(par[["input"]])
@@ -85,7 +94,8 @@ model <- CytoNorm::CytoNorm.train(
     transformList = NULL,
     normParams = list(nQ = par[["n_quantiles"]], goal = "mean"),
     seed = 42,
-    verbose = FALSE
+    verbose = FALSE,
+    recompute = TRUE
 )
 
 cat("Normalising using Cytonorm model trained using aggregates\n")